requirements.txt
langchain==0.0.267
langchain-core==0.1.4
langchain-openai==0.0.2.post1
langgraph==0.0.20
openai==0.28.1
python-dotenv==1.0.0
adbutils==2.8.0
google-generativeai==0.8.4
pillow==10.0.0
pydantic==1.10.8
adbutils==2.8.0
annotated-types==0.7.0
anyio==4.8.0
apkutils2==1.0.0
async-timeout==4.0.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
cigam==0.0.3
decorator==5.2.1
deprecation==2.1.0
distro==1.9.0
exceptiongroup==1.2.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.2
google-api-python-client==2.163.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-genai==1.5.0
google-generativeai==0.8.4
googleapis-common-protos==1.69.1
greenlet==3.1.1
grpcio==1.71.0
grpcio-status==1.71.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
idna==3.10
jiter==0.9.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.20
langchain-core==0.3.44
langchain-openai==0.3.8
langchain-text-splitters==0.3.6
langgraph==0.3.7
langgraph-checkpoint==2.0.19
langgraph-prebuilt==0.1.2
langgraph-sdk==0.1.56
langgraph-supervisor==0.0.9
langsmith==0.3.13
lxml==5.3.1
msgpack==1.1.0
openai==1.66.2
orjson==3.10.15
packaging==24.2
pillow==11.1.0
proto-plus==1.26.1
protobuf==5.29.3
py==1.11.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.10.6
pydantic_core==2.27.2
pyelftools==0.32
pyparsing==3.2.1
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
retry==0.9.2
rsa==4.9
sniffio==1.3.1
SQLAlchemy==2.0.39
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typing_extensions==4.12.2
uiautomator2==3.2.8
uritemplate==4.1.1
urllib3==2.3.0
websockets==14.2
xmltodict==0.14.2
zstandard==0.23.0
